[fl_dynload] Allow overriding low-level module loader
Before this patch, `Fl_dynload.load_packages` will unconditionally
call `Dynlink.loadfile` to load each particular compilation unit.

This works fine in almost all our scenarios, however when compiling
Web Workers with JSOO, we have found that `Dynlink.loadfile` takes a
very long time as it will compile `.cma` files to `.js`.

In the case of jsCoq and `coq-lsp` this compilation process is too
heavy. Prior to us using `findlib`, we used to trap all calls from Coq
to `Dynlink.loadfile`, and we used a pre-compiled `.js` version which
was much much faster.

This patch makes it possible for users of `Fl_dynload.load_packages`
to override the call to `Dynlink.loadfile` as to for example implement
the above cache setup, or to do other kinds of instrumentations.