Fail early in case either sh or sed are broken
'find -exec cmd {} ;' will call cmd for each file, however if the command has a non-zero exit code find will simply ignore the error code and continue as if nothing happened. Piping into read is much simply at the cost of not supporting filenames containing \n but ocamlfind doesn't have any and extremely unlikely thing to happen so this is safe to do.